Add method to get available position on the board
To start off, we'd need to randomly place two tiles on the board.
However, we don't know which slots on the board are available.
Hence, this method finds an empty slot on the board.

If its found, it will return its coordinates.
If not (board is full), it will return false.